pyproject.toml
[tool.poetry]
name = "testdjangooscar"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-oscar = "^3.2.4"
django-widget-tweaks = "^1.5.0"
treebeard = "^0.1.1"
sorl-thumbnail = "^12.10.0"
django-tables2 = "2.3"
click = "7.0.0"
django-haystack = "^3.2.1"
easy-thumbnails = "^2.8.5"
install = "^1.3.5"
django-oscar-api = "^3.2.4"
djangorestframework = "^3.14.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
